+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Removed
+
+- Code needed for VoterRegistry address and ABI changes. Reward epochs before the transition (417 on Flare and Songbird, 5451 on Coston, 5339 on Coston2) are no longer supported.
+
 ## [v1.3.0](https://github.com/flare-foundation/fdc-client/tree/v1.3.0) - 2026-7-14
 
 ### Changed